+++ packages/utils/test/mergeDefaultsWithFormData.test.ts
@@ -29,6 +29,13 @@ describe('mergeDefaultsWithFormData()', () => {
     expect(mergeDefaultsWithFormData({}, null, undefined, true)).toEqual({});
   });
 
+  it('should return null if default is null and formData is undefined and defaultSupercedesUndefined is true', () => {
+    const defaultValue = null;
+    const formData = undefined;
+    const defaultSupercedesUndefined = true;
+    expect(mergeDefaultsWithFormData(defaultValue, formData, undefined, defaultSupercedesUndefined)).toBeNull();
+  });
+
   it('should return undefined when formData is undefined', () => {
     expect(mergeDefaultsWithFormData(undefined, undefined)).toBeUndefined();
   });